Implementado verificação de campos na tela LIVRO e CLIENTES, seleção do Tipo de ENDERECO e Cadastro de Novo ENDERECO.
Passado para o controller SECAO a verificação de campo numérico.

+++ LivrariaDigital_teste/src/controller/OrdenaLivro.java
@@ -96,7 +96,7 @@ public class OrdenaLivro implements Comparator<Livro> {
 		@Override
 		public int compare(Livro primeiro, Livro segundo) {
 
-			return Float.toString( primeiro.getPrecoCusto() ).compareTo(Float.toString( segundo.getPrecoCusto() ));
+			return Double.toString( primeiro.getPrecoCusto() ).compareTo(Double.toString( segundo.getPrecoCusto() ));
 		}
 	}
 	
@@ -105,7 +105,7 @@ public class OrdenaLivro implements Comparator<Livro> {
 		@Override
 		public int compare(Livro primeiro, Livro segundo) {
 
-			return Float.toString( primeiro.getPrecoVenda() ).compareTo(Float.toString( segundo.getPrecoVenda() ));
+			return Double.toString( primeiro.getPrecoVenda() ).compareTo(Double.toString( segundo.getPrecoVenda() ));
 		}
 	}
 }
